feat: switch naive percentages to better formula
Idea: instead of performing the division right away on the percentage
number, multiply it by the amount first then perform the division.
Given that the amounts we manipulate are quite large in general (because
of the 1e18 decimal of ether and most ERC20 tokens), the rounding (if
it happens) will be minimal.

Inspiration:
https://forum.openzeppelin.com/t/fixed-point-arithmetic-support/356/3
https://www.youtube.com/watch?v=nsf46dzgCog